perf: 移除不必要的类型转换

+++ src/main/java/io/github/toggery/jt808/codec/B8103Codec.java
@@ -73,7 +73,7 @@ public final class B8103Codec implements Codec<B8103> {
             if (param != null) {
                 param.decode(version, fieldBuf, target);
             } else if (other == null || !other.decode(id, version, fieldBuf, target)) {
-                target.putUnknownParam((int) id, ByteBufUtil.hexDump(fieldBuf));
+                target.putUnknownParam(id, ByteBufUtil.hexDump(fieldBuf));
             }
         }
     }
